Add files via upload
Iterative : 
Uses a loop to repeatedly divide the search space in half until the target element is found or the entire array is searched.
Requires less memory than the recursive approach as it does not rely on function calls.
More efficient for large datasets due to its lower memory overhead.

Recursive :
Calls itself repeatedly with a smaller sub-array until the target element is found or the sub-array becomes empty.
Requires additional memory for storing function call stack frames.
Less efficient for large datasets due to its higher memory usage.

Time Complexity :
Both iterative and recursive binary search have a time complexity of O(log n), where n is the number of elements in the array. This means the search time increases logarithmically with the size of the array, making it efficient for searching large datasets.

Sequential Search :
Sequential search, also known as linear search, involves iterating through each element in the array sequentially until the target element is found. It has a time complexity of O(n), which means the search time increases linearly with the size of the array. This makes it less efficient than binary search for large datasets.