Add files via upload
Huffman Coding :
Huffman Coding is a lossless data compression technique that assigns variable-length codes to characters based on their frequencies. This means frequently occurring characters are assigned shorter codes, while less frequent characters are assigned longer codes. This approach allows for efficient data compression as it utilizes shorter codes for more frequently used information.

Greedy Approach :
Huffman Coding utilizes a greedy approach to assign codes to characters. This involves selecting the two characters with the lowest frequencies and merging them into a single node in a binary tree. The frequency of this new node is the sum of the frequencies of the two merged characters. The process repeats iteratively, merging the two nodes with the lowest frequencies until there is only one node remaining.